Overhauled Visual Studio Project
* Removed superfluous configurations
* Removed superfluous include directories and hard-coded library locations
* There are now only 3 configurations: Debug, Playtest and Release
* Debug and Playtest have the PLAYTEST and BALANCE_ENABLED preprocessor commands, while release doesn't
* Release has maximum optimisations enabled. Debug has none, and playtest has minimal
* Removed hard-coded output directories based on half-life installations, dlls are output in the solution folder (and user can customise as they see fit)
* All projects for all configurations compile without warnings or errors

+++ main/source/util/Stacktrace.cpp
@@ -299,7 +299,7 @@ char *stackTrace( int skipAssert ) {
 							unsigned char *ptr = (unsigned char *)*a;
 							if( ptr >= (unsigned char *)baseAddress ) {
 								for( int k=0; k<3; k++, ptr+=16 ) {
-									OUTS "  %X: ", ptr );
+									OUTS "  %X: ", (unsigned int)ptr );
 									for( l=0; l<16; l++ ) OUTS "%02X ", ptr[l] );
 									OUTS "  " );
 									for( l=0; l<16; l++ ) OUTS "%c", ptr[l]>=31&&ptr[l]<127?ptr[l]:'.' );
